refactor: separate movement logic from SelectionSystem
